requirements.txt
tensorflow==2.13.0
flask==3.0.3
numpy==1.24.3
pandas==2.0.3
pillow==10.4.0
xgboost==2.1.3